Moved parse_ansi functions to ansi.rs (where they probably should have been)
